chore: updates index.ts and Payload version
